@@ -1,6 +1,6 @@
 {
   "name": "tauri",
-  "version": "0.4.5",
+  "version": "0.5.0",
   "description": "Multi-binding collection of libraries and templates for building Tauri apps",
   "bin": {
     "tauri": "./bin/tauri.js"
@@ -57,17 +57,20 @@
     "imagemin-zopfli": "6.0.0",
     "is-png": "2.0.0",
     "isbinaryfile": "4.0.4",
-    "jsdom": "16.2.0",
+    "jsdom": "16.2.1",
     "lodash": "4.17.15",
-    "minimist": "1.2.0",
+    "minimist": "1.2.4",
     "ms": "2.1.2",
     "png2icons": "2.0.1",
     "read-chunk": "3.2.0",
-    "sharp": "0.24.1",
+    "sharp": "0.25.1",
     "webpack-merge": "4.2.2",
     "webpack-shell-plugin": "0.5.0"
   },
   "devDependencies": {
+    "@babel/core": "7.8.7",
+    "@babel/preset-env": "7.8.7",
+    "@babel/preset-typescript": "7.8.3",
     "@types/cross-spawn": "6.0.1",
     "@types/fs-extra": "8.1.0",
     "@types/imagemin": "7.0.0",
@@ -77,8 +80,9 @@
     "@types/ms": "0.7.31",
     "@types/sharp": "0.24.0",
     "@types/webpack-merge": "4.1.5",
-    "@typescript-eslint/eslint-plugin": "2.21.0",
-    "@typescript-eslint/parser": "2.21.0",
+    "@typescript-eslint/eslint-plugin": "2.23.0",
+    "@typescript-eslint/parser": "2.23.0",
+    "babel-jest": "25.1.0",
     "dotenv": "8.2.0",
     "eslint": "6.8.0",
     "eslint-config-standard-with-typescript": "14.0.0",
@@ -89,15 +93,16 @@
     "eslint-plugin-security": "1.4.0",
     "eslint-plugin-standard": "4.0.1",
     "husky": "4.2.3",
+    "is-running": "2.1.0",
     "jest": "25.1.0",
     "jest-mock-process": "1.3.2",
     "lint-staged": "10.0.8",
-    "lockfile-lint": "4.0.0",
-    "promise": "8.0.3",
+    "lockfile-lint": "4.1.0",
+    "promise": "8.1.0",
     "raw-loader": "4.0.0",
     "ts-loader": "6.2.1",
     "typescript": "3.8.3",
-    "webpack": "4.41.6",
+    "webpack": "4.42.0",
     "webpack-cli": "3.3.11",
     "webpack-node-externals": "1.7.2"
   },
