@@ -2,7 +2,7 @@ const cache = {}
 let initialized = false
 
 const proxy = new Proxy({
-  __consume () {
+  __consume() {
     for (const key in cache) {
       if (key in window.tauri) {
         const queue = cache[key]
@@ -21,10 +21,11 @@ const proxy = new Proxy({
       }
     }
     initialized = true
-  }
+  },
+  Dir: require('./fs/dir').Dir
 }, {
-  get (obj, prop) {
-    if (prop === '__consume') {
+  get(obj, prop) {
+    if (prop === '__consume' || prop === 'Dir') {
       return obj[prop]
     }
 
