@@ -2,9 +2,10 @@ const parseArgs = require('minimist')
 
 const argv = parseArgs(process.argv.slice(2), {
   alias: {
-    h: 'help'
+    h: 'help',
+    e: 'exit-on-panic'
   },
-  boolean: ['h']
+  boolean: ['h', 'e']
 })
 
 if (argv.help) {
@@ -21,4 +22,8 @@ if (argv.help) {
 
 const dev = require('../dist/api/dev')
 
-dev()
+dev({
+  ctx: {
+    exitOnPanic: argv['exit-on-panic']
+  }
+})
