@@ -19,7 +19,8 @@ export const spawn = (
   // TODO: move to execa?
   const runner = crossSpawn(cmd, params, {
     stdio: 'inherit',
-    cwd
+    cwd,
+    env: process.env
   })
 
   runner.on('close', code => {
@@ -29,7 +30,8 @@ export const spawn = (
       log(`Command "${cmd}" failed with exit code: ${code}`)
     }
 
-    onClose?.(code)
+    // eslint-disable-next-line @typescript-eslint/prefer-optional-chain
+    onClose && onClose(code)
   })
 
   return runner.pid
@@ -42,7 +44,7 @@ export const spawnSync = (
   cmd: string,
   params: string[],
   cwd: string,
-  onFail: () => void
+  onFail?: () => void
 ): void => {
   log(`[sync] Running "${cmd} ${params.join(' ')}"`)
   log()
@@ -60,7 +62,8 @@ export const spawnSync = (
     if (runner.status === null) {
       warn(`⚠️  Please globally install "${cmd}"`)
     }
-    onFail?.()
+    // eslint-disable-next-line @typescript-eslint/prefer-optional-chain
+    onFail && onFail()
     process.exit(1)
   }
 }
