@@ -7,8 +7,9 @@ module.exports = {
     'api/dev': './src/api/dev.ts',
     'api/init': './src/api/init.ts',
     'api/tauricon': './src/api/tauricon.ts',
+    'api/info': './src/api/info.ts',
     'helpers/tauri-config': './src/helpers/tauri-config.ts',
-    'api/info': './src/api/info.ts'
+    'helpers/spawn': './src/helpers/spawn.ts'
   },
   mode: process.env.NODE_ENV || 'development',
   devtool: 'source-map',
@@ -18,6 +19,10 @@ module.exports = {
         test: /\.tsx?$/,
         use: 'ts-loader',
         exclude: /node_modules/
+      },
+      {
+        test: /templates[\\/](tauri|mutation-observer)\.js/,
+        use: 'raw-loader'
       }
     ]
   },
