@@ -1,14 +1,18 @@
 mod cmd;
+mod salt;
+#[allow(dead_code)]
+mod file_system;
+mod dialog;
 
 #[cfg(not(any(feature = "dev-server", feature = "embedded-server")))]
 use std::path::PathBuf;
 use web_view::WebView;
 
 #[allow(unused_variables)]
-pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> crate::Result<bool> {
+pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> crate::Result<()> {
   use cmd::Cmd::*;
   match serde_json::from_str(arg) {
-    Err(_) => Ok(false),
+    Err(e) => Err(crate::Error::from(e.to_string())),
     Ok(command) => {
       match command {
         Init {} => {
@@ -20,48 +24,91 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
             event_init = event_init
           ))?;
         }
-        #[cfg(any(feature = "all-api", feature = "readTextFile"))]
+        #[cfg(any(feature = "all-api", feature = "read-text-file"))]
         ReadTextFile {
           path,
+          options,
           callback,
           error,
         } => {
-          crate::file_system::read_text_file(webview, path, callback, error);
+          file_system::read_text_file(webview, path, options, callback, error);
         }
-        #[cfg(any(feature = "all-api", feature = "readBinaryFile"))]
+        #[cfg(any(feature = "all-api", feature = "read-binary-file"))]
         ReadBinaryFile {
           path,
+          options,
           callback,
           error,
         } => {
-          crate::file_system::read_binary_file(webview, path, callback, error);
+          file_system::read_binary_file(webview, path, options, callback, error);
         }
-        #[cfg(any(feature = "all-api", feature = "writeFile"))]
+        #[cfg(any(feature = "all-api", feature = "write-file"))]
         WriteFile {
           file,
           contents,
+          options,
           callback,
           error,
         } => {
-          crate::file_system::write_file(webview, file, contents, callback, error);
+          file_system::write_file(webview, file, contents, options, callback, error);
         }
-        #[cfg(any(feature = "all-api", feature = "listDirs"))]
-        ListDirs {
+        #[cfg(any(feature = "all-api", feature = "read-dir"))]
+        ReadDir {
           path,
+          options,
           callback,
           error,
         } => {
-          crate::file_system::list_dirs(webview, path, callback, error);
+          file_system::read_dir(webview, path, options, callback, error);
         }
-        #[cfg(any(feature = "all-api", feature = "listFiles"))]
-        ListFiles {
+        #[cfg(any(feature = "all-api", feature = "copy-file"))]
+        CopyFile {
+          source,
+          destination,
+          options,
+          callback,
+          error,
+        } => {
+          file_system::copy_file(webview, source, destination, options, callback, error);
+        }
+        #[cfg(any(feature = "all-api", feature = "create-dir"))]
+        CreateDir {
           path,
+          options,
           callback,
           error,
         } => {
-          crate::file_system::list(webview, path, callback, error);
+          file_system::create_dir(webview, path, options, callback, error);
         }
-        #[cfg(any(feature = "all-api", feature = "setTitle"))]
+        #[cfg(any(feature = "all-api", feature = "remove-dir"))]
+        RemoveDir {
+          path,
+          options,
+          callback,
+          error,
+        } => {
+          file_system::remove_dir(webview, path, options, callback, error);
+        }
+        #[cfg(any(feature = "all-api", feature = "remove-file"))]
+        RemoveFile {
+          path,
+          options,
+          callback,
+          error,
+        } => {
+          file_system::remove_file(webview, path, options, callback, error);
+        }
+        #[cfg(any(feature = "all-api", feature = "rename-file"))]
+        RenameFile {
+          old_path,
+          new_path,
+          options,
+          callback,
+          error,
+        } => {
+          file_system::rename_file(webview, old_path, new_path, options, callback, error);
+        }
+        #[cfg(any(feature = "all-api", feature = "set-title"))]
         SetTitle { title } => {
           webview.set_title(&title)?;
         }
@@ -83,7 +130,7 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           callback,
           error,
         } => {
-          crate::salt::validate(webview, salt, callback, error);
+          salt::validate(webview, salt, callback, error);
         }
         #[cfg(any(feature = "all-api", feature = "event"))]
         Listen {
@@ -98,6 +145,22 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
         Emit { event, payload } => {
           crate::event::on_event(event, payload);
         }
+        #[cfg(any(feature = "all-api", feature = "open-dialog"))]
+        OpenDialog {
+          options,
+          callback,
+          error
+        } => {
+          dialog::open(webview, options, callback, error);
+        }
+        #[cfg(any(feature = "all-api", feature = "save-dialog"))]
+        SaveDialog {
+          options,
+          callback,
+          error,
+        } => {
+          dialog::save(webview, options, callback, error);
+        }
         #[cfg(not(any(feature = "dev-server", feature = "embedded-server")))]
         LoadAsset {
           asset,
@@ -108,7 +171,7 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           load_asset(webview, asset, asset_type, callback, error)?;
         }
       }
-      Ok(true)
+      Ok(())
     }
   }
 }
